requirements.txt
# Core FastAPI
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pymongo==4.5.0
pydantic==2.12.4
python-dotenv==1.2.1
python-multipart==0.0.20
bcrypt==4.1.3
PyJWT==2.10.1
python-jose==3.5.0
openai==1.99.9
litellm==1.79.3
pillow==12.0.0
pandas==2.3.3
openpyxl==3.1.5
requests==2.32.5
beautifulsoup4==4.14.2
firebase_admin==7.1.0
APScheduler==3.11.1
stripe==13.2.0
email-validator==2.3.0
passlib==1.7.4
starlette==0.37.2
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
